Fix diff and merge tests
We do a dumb merge right now that overwrites everything. We can define a
smarter merger and pass as argument as desired.

We should also put the string sorting inside the autogenerated code, I
think, instead of outside. It will help be consistent all accross. We
can also push the hash generation to the autogenerated code.