Create method to update context, for now (#2181)
Note: Currently this method is used to swap between contexts in test cases. It is not a method that should be relied
on long-term as we have methods that expect an unchanged context and they will be angry.

However, this allows us to test authz, so it'll do for the moment.