Install latest available version of NumPy (#2455)
This change is prompted by the freshly-released pandas==1.3.0 rejecting numpy<1.17.3.

The reason me-from-2-years-ago put numpy<1.17 in our requirements was because 1.16 is the last series to support Python 2. Really, we ought to just grab the latest available NumPy for whatever Python version is being run, rather than stick with a ~2-year-old version.