fix: fix build by reenabling skipTests
CI build was setting `-Dmaven.skip.test=true`, which would not only not run tests but would also not even build them. 
It was recently changed to use `-DskipTests`, however this pom.xml strangely overrides that setting. Applying the setting in the pom is unexpected.
This change removes the override returning to the previous state of not running tests in CI.  We should look at enable tests in CI.