feature: Have client use url from backend when available
The full prediction url was added to the API response when fetching the details of an endpoint.  However, this value may not always be present.  This change expands the get_deployed_model method with a call to _get_json_by_id (where the newly added "full_url" value resides) so it can be passed along when it exists.  If it does not exist, the default behavior remains the same and the url is assembled.