feature: [VRD-535] Have client use endpoint url from backend when available. (#3290)
* feature: Have client use url from backend when available

The full prediction url was added to the API response when fetching the details of an endpoint.  However, this value may not always be present.  This change expands the get_deployed_model method with a call to _get_json_by_id (where the newly added "full_url" value resides) so it can be passed along when it exists.  If it does not exist, the default behavior remains the same and the url is assembled.

* test: Add unit tests to cover additiona of full_url to API schema

The patching technique used here is a bit ugly.  I'd like to do something more sophisticated in the future.
This change adds 3 unite tests:
* Verify that the get_deployed_model calls the other methods it uses correctly
* Verify that the full_url field is returned when it is present in the API response
* Verify that the full_url is built and returned when it is NOT in the API response