Update Readme.md: Remove "How to Contribute" line not in use
I noticed that the mentioned line  "How to Contribute" in "What's In This Document" lacked any description or content for the "How to Contribute" section. Consequently, it appeared superfluous and could potentially confuse or mislead contributors. To address this issue, I have made the necessary changes to the Readme file to remove the line entirely.
Additionally, it ensures that the Readme accurately reflects the available information and helps new contributors get started on the right track.