Feature: Add OpenTelemetry AutoConfigured SDK initialisation to TracingExtension
The TracingExtension.java class was extended by introducing a new method
to initialize an OpenTelemetrySdk instance using
AutoConfiguredOpenTelemetrySdk builder. This instance sets the global
OpenTelemetry SDK, and configures various properties including tracing
exporter, propagators, and disabled resource providers.

The purpose of this change is to streamline the handling of tracing in
testing environment. Developers can now easily initialise the Open
Telemetry SDK, which may be useful for enhanced tracking of testing
processes.

Also, the pom.xml file in the common-testing directory was updated to
include a dependency on 'opentelemetry-sdk-extension-autoconfigure' for
enabling the AutoConfiguredOpenTelemetrySdk usage.