Add validation and tests for ExperimentRun's name update
Added validation in FutureExperimentRunServiceImpl to ensure that the
experiment run's name is not blank. An exception is thrown if a name
argument isn't present when updating the experiment run's name. To
ensure the proper implementation, added corresponding tests in
ExperimentRunTest to verify correct results and the absence of any
exceptions when the name is properly updated. Also tested the scenario
when the name argument is missing.