Add versioning in ExperimentRun's name update
Incremented the version number each time the name of the experiment run
is updated. This enhancement was done to track the changes in the name
over time. Removed unneeded methods `updateModifiedTimestampV2` and
`updateVersionNumberV2` as the changes are handled within the `UPDATE`
sql statement. Verified these changes in the `ExperimentRunTest` by
asserting that the version number increases by one after a name update.