Feature: Implement experiment run name update functionality (#4098)
* Implement experiment run name update functionality

Added functionality to update the name of an experiment run by modifying the methods in FutureExperimentRunServiceImpl and FutureExperimentRunDAO classes. The changes permit the updating of the experiment run name in the database. This functionality was implemented in response to the need for modification of experiment run details after their creation. Also, optimized the import statements in FutureExperimentRunDAO class for better readability and maintenance.

* Update experiment run name functionality

Implemented new methods updateModifiedTimestampV2 and
updateVersionNumberV2 in FutureExperimentRunDAO for update operation of
experiment run name. Refactored updateExperimentRunName method to use
the new implementations. Optimal import statements also added as part of
readability improvement. Aim is to provide efficient way of updating
experiment run name after creation.

* Add validation and tests for ExperimentRun's name update

Added validation in FutureExperimentRunServiceImpl to ensure that the
experiment run's name is not blank. An exception is thrown if a name
argument isn't present when updating the experiment run's name. To
ensure the proper implementation, added corresponding tests in
ExperimentRunTest to verify correct results and the absence of any
exceptions when the name is properly updated. Also tested the scenario
when the name argument is missing.

* Add versioning in ExperimentRun's name update

Incremented the version number each time the name of the experiment run
is updated. This enhancement was done to track the changes in the name
over time. Removed unneeded methods `updateModifiedTimestampV2` and
`updateVersionNumberV2` as the changes are handled within the `UPDATE`
sql statement. Verified these changes in the `ExperimentRunTest` by
asserting that the version number increases by one after a name update.

* removed deprecated